+++ stdlib_strtod.c
@@ -326,7 +326,7 @@ strtod(const char *str, char ** ptr)
 					double divisor;
 
 					/* A negative exponent means division. */
-					divisor = pow(radix,(double)exponent);
+					divisor = pow((double)radix,(double)exponent);
 					if(divisor != 0.0)
 					{
 						new_sum = sum / divisor;
@@ -343,7 +343,7 @@ strtod(const char *str, char ** ptr)
 				else
 				{
 					/* A positive exponent means multiplication. */
-					new_sum = sum * pow(radix,(double)exponent);
+					new_sum = sum * pow((double)radix,(double)exponent);
 					if(new_sum < sum)
 						error = ERANGE;
 					else
